test_SurfStatF.py working again
